requirements.txt
Flask
gunicorn
Pillow
pdf2image
reportlab
pypandoc
python-docx
pikepdf
img2pdf
